+++ .gitignore
@@ -3,3 +3,4 @@ iso-chooser-menu
 out.vars
 /*.json
 test/runtests
+/.tags
