+++ args.h
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2022-2023 Canonical Ltd.
+ *
+ * SPDX-License-Identifier: GPL-3.0
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3, as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranties of
+ * MERCHANTABILITY, SATISFACTORY QUALITY, or FITNESS FOR A PARTICULAR
+ * PURPOSE.  See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see
+ * <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+typedef struct _args_t
+{
+    char *outfile;
+    int  num_infiles;
+    char **infiles;
+} args_t;
+
+args_t *args_create(int argc, char **argv);
+void args_free(args_t *args);
